requirements.txt
PyYAML==5.4.1
jsonschema==3.2.0
xmltodict==0.12.0
requests==2.25.1
urllib3==1.26.2
google-cloud-monitoring==2.2.1

